Move wxAnimationDecoderList-related methods to wxAnimation
