+++ interface/wx/animate.h
@@ -438,15 +438,6 @@ public:
        @c wxAnimation and @c wxAnimationCtrl.
      */
     virtual wxColour GetBackgroundColour() const;
-
-
-    static inline wxAnimationDecoderList& GetHandlers();
-    static void AddHandler(wxAnimationDecoder *handler);
-    static void InsertHandler(wxAnimationDecoder *handler);
-    static const wxAnimationDecoder *FindHandler( wxAnimationType animType );
-
-    static void CleanUpHandlers();
-    static void InitStandardHandlers();
 };
 
 
@@ -480,6 +471,15 @@ public:
 
     bool LoadFile(const wxString& name, wxAnimationType type = wxANIMATION_TYPE_ANY);
     bool Load(wxInputStream& stream, wxAnimationType type = wxANIMATION_TYPE_ANY);
+
+    
+    static inline wxAnimationDecoderList& GetHandlers();
+    static void AddHandler(wxAnimationDecoder *handler);
+    static void InsertHandler(wxAnimationDecoder *handler);
+    static const wxAnimationDecoder *FindHandler( wxAnimationType animType );
+
+    static void CleanUpHandlers();
+    static void InitStandardHandlers();    
 };
